Separated lib and prompt filterq
This way the repo can also be used as a dependency in cmder
where there is some extra stuff to be done before it can be added
to the prompt
